mlops/requirements.txt
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
dvc==3.50.1
pyyaml==6.0.1
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
